--- yii/bootstrap/enum/ProgressEnum.php
@@ -1,19 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @link http://www.yiiframework.com/
- * @license http://www.yiiframework.com/license/
- */
-
-namespace yii\bootstrap\enum;
-
-/**
- * ProgressEnum provides easy access to all predefined progress bar set of named values
- *
- * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @since 2.0
- */
-class ProgressEnum
-{
-
-}
